description
This project deals with detecting spam messages with high accuracy by using ML classification algorithms .
Here Countvectorizer is used for extracting features form a given dataset and design model by generating tests and training sets from given data. Then the naive bayes classifier is applied for clustering and the test and training set is given as input based on this data given message is tested if it is spam or not. The other classifiers used for the spam message classification are:
1.Decision tree
2.Random Forest
3.KNN
4.Bagging
5.AdaBoost