Update README.md
Detection of spam is important for securing message and e-mail communication. The accurate detection of spam is a big issue, and many detection methods have been proposed by various researchers. However, these methods have a lack of capability to detect the spam accurately and efficiently. To solve this issue, we have proposed a method for spam detection using machine learning predictive models. The method is applied for the purpose of detection of spam. The experimental results obtained show that the proposed method has a high capability to detect spam. The proposed method achieved 98% accuracy which is high as compared with the other existing methods. Thus, the results suggest that the proposed method is more reliable for accurate and on-time detection of spam, and it will secure the communication systems of messages and e-mails.

HOW IT WORKS?
 Countvectorizer is used for extracting features form a given dataset and design model by generating tests and training sets from given data. Then the naive bayes classifier is applied for clustering and the test and training set is given as input based on this data given message is tested if it is spam or not. The other classifiers used for the spam message classification are:
1.Decision tree
2.Random Forest
3.KNN
4.Bagging
5.AdaBoost